refactor: move execution watcher logic to a separate module
